Implemented map printing.
Printing a map displays the basemap, clusters and markers, and popups.

If the filter popover is displayed, the filters are listed in a panel
over the map.

If there are directions, they are printed in a textbox below the map.

Many controls are hidden during printing of the map, such as:

* 'Find Me' button
* 'Search' button
* 'Print' button
* 'About' link
* 'FAQ' link
* Any other UI buttons in InfoWindows, popups, or direction panes