Move v1 frontend to _v1 temporarily
We're completely redoing the frontend, but we're going to want to be
able to steal code from the old one.  I'm temporarily moving the old one
to *_v1 so we can conveniently look at the old code while developing
the new stuff.  Eventually once we have the next version up and running
we should delete *_v1.