Move localemiddleware so that it is used first
This fixes the 'caching' problem that was causing stickiness
of languages on the browse page when zooming in and out, etc.

I'm not entirely sure how this works, but according to
https://docs.djangoproject.com/en/dev/topics/cache/#order-of-middleware-classes
middleware is loaded in reverse order, so the last one is loaded first,
moving locale middleware to the end of the list seems to solve the
issue on my deployment (with nginx, etc.). This will require
further testing on the staging server