Remove language prefix from faq page
This fixes the faq page not reloading on language change. The
original problem originated because the change language url
sets the language in the cookie, then reloads the url it came
from. In the original case, this meant it was reloading the
url with the language information which was how language
was determined.

Turns out, the logic in the faq view doesn't require that
prefix and works fine just looking at the language header
in the request.