Add functionality to check if lat+lng in url is geolocated or not
This fixes a problem where the user browses back to the page with
a url with lat+lng, but that lat+lng does not represent a geolocated
position. History.js sets a data object in history that is checked
when loading a page. If this data object is present and not undefined,
we know that that the point in the url does not represent a geolocated
position.