Move logic to set top margin for list view into browse.js
This fixes the problem where the results list was sometimes
being hidden by the filters. The problem seemed to be from
javascript functions/events being fired at random times.
This fixes the problem by calling this function only during
the actual construction of list results when we know the
filter div is already populated.